requirements-dev.txt
# Development dependencies for Fugatto Audio Lab
# Install with: pip install -r requirements-dev.txt

# Include production dependencies
-r requirements.txt

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-xdist>=3.3.0

# Code Quality
black>=23.7.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.5.0
ruff>=0.0.280

# Pre-commit
pre-commit>=3.3.0

# Evaluation Tools
pesq>=0.0.4
pystoi>=0.3.3
torchmetrics[audio]>=1.0.0
crowdkit>=1.2.0

# Development Tools
bumpversion>=0.6.0
twine>=4.0.0
wheel>=0.40.0